created front end dashboard structure
