+++ src/components/dashboard/contentarea.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Contentarea() {
+  return (
+    <div className='basis-10/12 border-slate-800 border-2'>
+      Contentarea
+    </div>
+  )
+}
+
+export default Contentarea
