+++ src/components/temporary folder/buyStocks.js
