+++ src/components/temporary folder/stocksquery.js
